MODIFY : Corrections on transmit functions

+++ libcanard/funccanard.h
@@ -12,7 +12,7 @@ uint8_t deserializeFrame(uint8_t *buffer, uint16_t* identifier, uint8_t buffer_s
 
 uint8_t insertStuffedBits (uint8_t* buffer_in, uint8_t* buffer_out, uint8_t payload_size);
 
-uint8_t removeStuffedBits(uint8_t* buffer_in, uint8_t* buffer_out, uint8_t buffer_size);
+uint8_t removeStuffedBits(uint8_t* buffer_in, uint8_t* buffer_out, int buffer_size);
 
 uint8_t pleaseTransmit(CanardTxQueueItem* ti, CanardTransferMetadata transfer_metadata);
